tests: migrate integration-cli stats tests to integration
